qca_cld: Add VOSS API to get monotonic boot time
Add VOSS API to get monotonic boot time for recording WMI command/event
timestamp.

CRs-fixed: 669914

+++ CORE/SERVICES/WMI/wmi_unified_priv.h
@@ -49,13 +49,13 @@ typedef adf_nbuf_t wmi_buf_t;
 struct wmi_command_debug{
 	u_int32_t command;
 	u_int32_t data[4]; /*16 bytes of WMI cmd excluding TLV and WMI headers*/
-	u_int32_t time;
+	u_int64_t time;
 };
 
 struct wmi_event_debug{
 	u_int32_t event;
 	u_int32_t data[4]; /*16 bytes of WMI event data excluding TLV header*/
-	u_int32_t time;
+	u_int64_t time;
 };
 
 #endif /*WMI_INTERFACE_EVENT_LOGGING*/
